Group all the "raw" implementations.
