+++ exec/io.go
@@ -64,91 +64,3 @@ type StdioSetter interface {
 	// standard error when the command starts.
 	StderrPipe() (io.ReadCloser, error)
 }
-
-// RawStdio exposes low-level machinery for managing stdio.
-type RawStdio interface {
-	// SetStdio sets the stdio to use.
-	SetStdio(values Stdio) error
-
-	// StdinPipe, StdoutPipe, and StderrPipe each create a pipe for
-	// the corresponding stdio stream.
-	StdinPipe() (io.WriteCloser, io.Reader, error)
-	StdoutPipe() (io.ReadCloser, io.Writer, error)
-	StderrPipe() (io.ReadCloser, io.Writer, error)
-}
-
-// CmdStdio is a basic StdioSetter implementation for a single set
-// of stdio streams.
-type CmdStdio struct {
-	// Raw provides the underlying functionality.
-	Raw RawStdio
-
-	// Stdio holds the Stdio information.
-	Stdio *Stdio
-}
-
-// SetStdio implements StdioSetter.
-func (s *CmdStdio) SetStdio(values Stdio) error {
-	stdio, err := s.Stdio.WithInitial(values)
-	if err != nil {
-		return errors.Trace(err)
-	}
-
-	if err := s.Raw.SetStdio(stdio); err != nil {
-		return errors.Trace(err)
-	}
-
-	s.Stdio = &stdio
-	return nil
-}
-
-// StdinPipe implements StdioSetter.
-func (s *CmdStdio) StdinPipe() (io.WriteCloser, error) {
-	w, r, err := s.Raw.StdinPipe()
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	if err := s.SetStdio(Stdio{In: r}); err != nil {
-		if err := w.Close(); err != nil {
-			logger.Errorf("while closing stdin pipe: %v", err)
-		}
-		return nil, errors.Trace(err)
-	}
-
-	return w, nil
-}
-
-// StdoutPipe implements StdioSetter.
-func (s *CmdStdio) StdoutPipe() (io.ReadCloser, error) {
-	r, w, err := s.Raw.StdoutPipe()
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	if err := s.SetStdio(Stdio{Out: w}); err != nil {
-		if err := r.Close(); err != nil {
-			logger.Errorf("while closing stdout pipe: %v", err)
-		}
-		return nil, errors.Trace(err)
-	}
-
-	return r, nil
-}
-
-// StderrPipe implements StdioSetter.
-func (s *CmdStdio) StderrPipe() (io.ReadCloser, error) {
-	r, w, err := s.Raw.StderrPipe()
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	if err := s.SetStdio(Stdio{Err: w}); err != nil {
-		if err := r.Close(); err != nil {
-			logger.Errorf("while closing stderr pipe: %v", err)
-		}
-		return nil, errors.Trace(err)
-	}
-
-	return r, nil
-}
